More refactoring, esp. error detection
Previously, error detection (and other messages to the development environment) were somewhat all over the place: different kinds of objects did error detections at different stages, and using different objects and methods.  Part of the problem was that reporting an error meant that you had access to a DevEnvironment instance, either because you kept one as a member or because it was a param to a method.

Now, everything (except [Expr]s) can report an error by delegating it to another object in a chain of objects.  TST and AST objects all keep a [Cell] object, which delegates that to the [Sheet] that they appear in, which delegate that to the [SheetProject] that they appear in, and that [SheetProject] keeps a DevEnvironment instance.  (When we're not actually in a project, like for some tests, you can just give them a DummyCell object that will ignore the error.)

The handy thing about this chain of delegation is that it means that ASTs, especially, don't have to keep around information about sheet names, positions, etc.  Some object up the chain knows that.  You just send the information you know, and by the time it gets to [SheetProject] all the information will be present.