Sheets re-integration
Hooked the new semantics up to the GSuite add-on.

One small change: the default symbol for each namespace is now called "__DEFAULT__" rather than "__MAIN__"; "__MAIN__" is now only used at the level of the project/root.  (We never actually refer to that default symbol by name, actually, so it doesn't really matter what we call it.  But we have to call it something so that it gets stored in the namespace.)