Re-integrated sampling
I had turned off sampling for a while, while I rewrote the basic search algorithm (i.e. breadth-first search through tapes).  Now that we're integrating with GSuite again, though, we have to turn on sampling again.

This commit implements a new sampling algorithm, that should hopefully be more efficient than the last one.  Note that it's depth-first as far as characters go, but still breadth-first on tapes (that is, it will still do derivs/deltas w.r.t. tapes in a cycle).