Refactoring of project objects
Although sheets and everything smaller than them are now factored out into several hierarchies, the Project object was a little bit of everything -- it represented a directory, a program, and an AST all at once.  This refactoring splits that up somewhat.  There's now a SheetProject object that represents the project as a collection of sheets, and is responsible for loading them from the dev env and creating a TST for them.  There's also a TstProject that's fairly simple, it really just knows how to create an AST of itself.  There's no project-equivalent in the AST or deriv hierarchies, though; it's just another namespace from that point-of-view.

While I was at it, I move the tests/ast tests in tests/, replacing the old tests that were there.  (They had, by then, taken over as the only real tests; I wasn't using the old ones anymore.)  It had gotten too cumbersome to update the old, unused tests to reflect all the high-level refactoring.