Header.toAST now creates correct cell references
Because the Header.toAST() methods used the convenience methods for AST construction (e.g. Seq, Uni, Lit), any AST created in this way wouldn't be able to send error/warning messages to the correct cell.  (The convenience methods all use DummyCell, which just ignores all interface messages.)  This commit fixes that, so that the AST's .cell member will now correctly point to the cell it's in.  It happens that none of our current error messages use this, but it's good for the future.

With this is a small refactoring of Headers.  Previously, all Headers had a member named .text, which for most Headers was only used in testing.  The exceptions were the AtomicHeaders, which used .text to determine what their .hue was, and that to answer a .getColor() call.  Rather than have all Headers implement .text and .hue, now only the AtomicHeaders do.  This simplifies the constructors and thus the parser/combinator grammar a bit, too -- we're no longer passing in otherwise-useless strings like "/" and "startswith" into every constructor.