Tests for cell parsing
Cell parsing is the process of turning a content cell with boolean content( e.g. ~(1SG|2SG)) into a tree. We weren't ever testing the cell parsing alone (only as part of sheets), so this adds some simple tests that the cell parser is doing its job.

This also adds some documentation for miniParser, since that was undocumented.