Re-integrate parsing of user-supplied inputs
For example, in the GSuite interface, you can supply a value to any field and it will only sample/generate outputs compatible with that field.  I had turned that off; now it's back on.

I also found and fixed a bug introduced by the interaction of the new generate algorithm and the way MatchState treats tapes.  (I'm pretty sure this is only a bug in the new semantics.)  The problem was that, when MatchState "cares about" tapes that aren't actually in its child, then the tapes created by collectVocab() are not the same as those returned by getTapes().  What happens then is that filters/joins can create dot-stars based on the results of getTapes(), but then those dot-stars are never dealt with because the generation algorithm uses the tapes from collectVocab.  So either MatchTapes.getTapes() has to not report those irrelevant tapes, or MatchTapes.collectVocab() has to create those tapes.

I did the first solution, but if later we find we need to do the second solution instead, is IS possible to create empty tapes; it's what you get when you only tokenize the empty string.