Re-integrating unit tests
Getting unit tests working again was very involved, because of (in the end) static name resolution.  So this update also comes with a lot of structural refactoring -- which is good, because it got rid of some strange classes like Root, which was both a facade and a namespace and... well, good riddance.

We used to just run unit tests at the point of the tree where they existed, and because we were doing dynamic name resolution, that worked any time.  But now we're doing static name resolution, and that only works if you run qualifyNames() from the root of the tree.  So it's not just enough (as we'd been doing) to grab a particular AST, get the Root object, and ask Root to generate from the corresponding expression.  (Because if it references symbols outside of that scope, they won't have been available during name qualification.)

Changes:

(1) All objects related to a grammar can now generate -- Sheet components, TSTs, ASTs, and Exprs.  You don't have to get a Root, there's no more Root class.  Sheets, TSTs, and ASTs allow you also to specify a symbol name and restrictions (i.e., what you want to parse).  (Exprs don't allow either because they don't know about namespaces -- they exist after name resolution -- and restrictions require tape information that only ASTs and above can know.)

(2) EmbedExprs no longer keep an INamespace, they just keep a dict of names and exprs.  All the former behavior of namespaces at the Expr level has disappeared, so we don't need a fullfledged interface/class anymore.

(3) For a few commits there, TstTables didn't have any TstRow children, they just kept a reference to the last TstHeadedCell in the row.  But it turns out we need a single piece of information from that row that's inconvenient to get from the last TstHeadedCell: the position of the FIRST cell in the row.  We need that to mark success/failure of unit tests in a consistent place (since the first cell of the row is often in the same column, whereas the last cell more often varies).

(4) ASTs, upon constructing Exprs, save them.  This is mostly so that, when we do ask for the expression of a sub-AST (like a named symbol or unit test), we don't obliterate names that we properly resolved when constructing expressions from higher up the tree.