Finished port of Replace to new semantics
Finished porting Darlene's implementation of phonological replacement rules to this branch, and the tests for them.

A few bugs/snags were uncovered in the process:

(1) RenameExpr.delta() should have been checking for the tape param to be equal to its .fromTape, just like we do in .deriv().  We don't, after all, want to attempt to discard "t1" inside a Rename(t1->t2) expression, but that's what happens if we just pass on the request to the child.  Instead, the correct thing to do is just return "this".

(2) MatchExpr.delta() wasn't quite right.  It was saying that you could only "move on" if all of the tapes inside the child were nullable, but there are situations where that isn't true.  For example, if you've matched "h" on t1, but are awaiting its match on t2, you have "h" in the t1 buffer but you also have something (e.g. h) inside the child at t1.  That's not nullable, but so long as you satisfy both the buffer and the child later (and nothing more than that), then t1 is nullable right then.  So now MatchExpr.delta() returns the intersection of the buffer and child.delta().  (Note that that's not a MatchExpr.  There's no more matching, you've stopped on one of the tapes and therefore there's no novel material possible on any of the other tapes, it's just a matter of handling the consequences of what you've already matched.)

(3) Getting MatchExpr.delta() correct when there's an irrelevant tape in the child (e.g. in our tests, a t3 when we're only matching t1 and t2) would be pretty convoluted -- possible, but convoluted, because you have to treat relevant and irrelevant content differently when constructing that intersection.  Given that we might not even use this possibility, I commented out the tests for that, for now at least.