Removed abandoned sub-repos
We were still carrying around the "version 1" language (the parser-combinator one) and packages that assumed that (like the Angular component), but we're now onto the third version of the language (the one based on Brzozowski derivatives), and it's a very different language.

With Fineen's port of the Michif gramble from v1, there are no longer any live projects using v1, so it's time to remove it.