Sampling without maximum "tries"
When it was possible to fail to sample (because of the bad old way of sampling we'd been doing), an empty result didn't necessarily mean there wasn't a possible output, it just meant we hadn't found one.  So we kept trying, up to a maximum number of tries.

Now that shouldn't happen -- if there are any possible outputs, we should find one.  If we get no results, then there weren't any.  So "maxTries" is no longer relevant; we shouldn't try any more # of times than the samples the client wants.  (Actually, we should abort completely at the first failure.  But we tend to call this with low number of samples anyway, like 5.)