Performance fixes/improvements
(1) Even though the algorithm itself was a generator (and so can stream results), our Edifice object ("Gramble") was a function, so interfaces couldn't stream results to (say) load up a results spreadsheet incrementally.  So now there's a method .generateStream for that.

(1.5) Also made a depth-first search generation algorithm, out of concern that having too large a search queue (for breadth-first search) was causing memory problems.  It probably wasn't, but a nice effect of the depth-first is that it ends up streaming results at a more even pace, rather than churning with no output for a long time and then returning a bunch all at the end.  The downside of depth-first is that it will act inappropriately w.r.t. recursion sometimes, but we're (temporarily) ignoring recursion in favor of more pressing issues right now anyway.

(2) Fineen's medium-sized grammar was somehow blowing the stack, when it really shouldn't.  It ended up being that the way intersections calculated their derivatives led to very large unions.  Since .deriv() returns the components of its (metaphorical) unions as separate yields, and we make a new intersection for each combination of such yields, intersection derivs create a lot of intersections.  Since these eventually do end up in a big union, the resulting union can be very large, large enough that (because it's a linked list) it can blow the stack when we ask for any function that recurses.  (Even .id!)

It's really more appropriate and performant here (and possibly everywhere) to make intersections of the .disjointDeriv() results instead.  That keeps compatible results (that is, results representing the same character outputs) together as a union, and thus there are many many fewer intersections that we have to make.

In the longer term, we need to go back to an array-based implementation of UnionExpr.  There really ARE unions that we'll have to traverse that are larger than the amount of stack we have available -- if a programmer wants a symbol to represent a 10k-word dictionary, they're within their rights to want that, and that WILL blow the stack if implemented as a linked list.  But for now I'm keeping it as a linked list; it's temporarily useful to have something go very wrong when any process creates unions of hundreds of elements, because (with the small-to-medium grammars we're currently working with), that's probably due to an implementation error.

Other small changes:

(1) Made a special DotStarExpr.  It's simpler to think about and reason about as a unit -- look at how simple those deltas and derivs are, compared to star, which is hard to reason about.  In practice it doesn't really speed things up, it's just simpler I think.

(2) Reimplemented an older way of doing filters, just in case something about our intersection-based implementation was causing the stack to blow.  That wasn't ultimately the problem, but this new (or rather, old) implementation does work, so I set it aside in derivsExperimental.ts for now.  In theory it'll be a little more performant (unlike the intersection there's no concatenation to iterate through) but I didn't notice a big difference.