Dedicated join/filter Exprs
In the previous commit, I tried out using a dedicated FilterExpr rather than construct them out of intersections/concats/dotstars.  (This is closer to what we used to do.)  But it didn't (at the time) seem to make a big time difference.

However, swapping the order of its operands (so that it tries child2, the expression expressing the filter, first), seems to reliably give it a performance boost over the intersection/concat/dotstar implemnentation.  (This is still true if we similarly flip the operands of the intersection.)  The improvement is maybe 20-30% on the Michif grammar.

The reason for this is probably that filters tend to have fewer alternations under them than the grammar they're filtering, so we end up going down fewer paths on the grammar side.  (E.g., rather than working out 100 possibilities for the next character on the grammar side, which may itself be a complex operation, and then checking them on the filter side, we work out the relatively few possibilities (often only 1) on the filter side, and then try to match it with the grammar.)

(Also, the committed version from last times seems to be missing the .disjointDeriv fix -- I had thought I'd done that for both implementations, but I guess not.  So that's fixed, too.)

So, in this commit I:

* implemented the disjointDeriv fix (mentioned in the last commit) for FilterExpr too.

* swapped the order of the operands, so that we loop over child2 (the filter) before child1 (the grammar)

* implemented a JoinExpr in the same manner.  (This is, again, pretty close to our old implementation of Joins.)

* refactored AstStartsWith, AstEndsWith, and AstContains so that they use the same .constructExpr() code as AstFilter.  They had previously been subclasses of AstJoin; now they're subclasses of AstFilter, and only differ in how they construct the filter expression.

Only a somewhat unrelated note, I made it so that AST's tapes are stored as lists rather than sets, because next I want to try iterating through tapes in a pre-defined order.  (Well, technically JavaScript Set objects do maintain their original order-of-insertion when you're iterating over them, we could have kept them as sets.  But I'm a bit hesitant to rely, here, on a "side" feature of sets, when other languages' Set objects don't have this property.  (Although we DO rely on it elsewhere...))