Rectification of names
It wasn't really accurate to call our "AST" objects as such; these are a semantic representation that's doesn't line up with the actual syntactic objects of our programming language.  The real ASTs of this programming language are the representations we call TSTs.

So this commit renames them to Grammars, which better represents what they actually are: they're the semantic representation of the grammar in the abstract, as opposed to its realization in programming language syntax OR in the exact symbolic objects juggled by the generation engine, and it's the level at which we check for semantic well-formedness.

It also renames the filename to "grammars.ts", and changes other names referring to ASTs (like TSTComponent.toAST() is now TSTComponent.toGrammar()).

Meanwhile, renamed the "deriv.ts" file to "exprs.ts" (because it's where Expr objects live) and "cellParser.ts" to "cells.ts" (by analogy with "headers")