Fixed tape priority orders
Before, we had an effectively random priority for the order in which we examined tapes, and then rotated between them (so as to not inadvertantly spend a lot of time generating from a low-priority tape -- i.e. one that's not going to fail -- before trying a high-priority tape).

This commit tests the idea that we should just be figuring out that priority.  This is a (fairly simple) attempt at that, which says that during the calculateTapes() stage, each grammar object should return the tapes in the priority it prefers.

For example, filters prioritize their 2nd child's tapes, joins prioritize the intersection of their children's tapes, sequences prioritize their earlier members' tapes, etc.

We could probably do this a little better, but this seems to be good enough for now.  It doubles the speed of generation for the Michif grammar compared to the last commit (which was already 20-30% faster than the commit before that).