Experiments with automatic multichar detection
This commit breaks a lot of things and invalidates a lot of tests, so I'm moving it to a new branch.

The problem it's trying to solve is that there are many tapes that aren't really representing strings (e.g. sequences of tokens).  For example, a "subject" field might have the possible values "1SG", "2SG", "3SG", "1PL", "2PL", "3PL", but never have a possible value "1SG2SG3SG" or "SSSSSS".  Really, those entries are one token, and there's never going to be a sequence of them, so treating them "1SG" as ["1", "S", "G"] is both inappropriate and inefficient.

Inefficient example:
  * If some entry has the field tense:"PRES.PROG", and we're filtering on tense:"PRES.PERF", and these are really atomic wholes rather than strings, then we should be able to fail right away, we don't have to wait until the sixth token to fail.

Inappropriate example:
  * If the possible values of UNIT are { "SPRING", "SUMMER", "FALL", and "WINTER" }, we expect the negation of "WINTER" to have three values.  It doesn't, it has an infinite number of values, like "WUNTER" and "LASSIE": it's every string made up of the letters in those values that aren't "WINTER".

This commit is a first stab at detecting which fields are atomic rather than strings.

This invalidates a number of tests: anything that counts vocabulary items (obviously), negation (it completely changes the results of negation, obviously), and certain match/replace (haven't really implemented this for match/replace yet, I expect those to fail).

It also reveals a more general issue about how namespaces work -- the question "are you all of your children or just the last" keeps rearing its ugly head.  We should quash this, it rests on a bit of a category error.

I also moved this to another branch because I'm not really certain this is something we want to do.  The problem is basically surprise -- this is the first instance I can think of where an optimization changes the actual outputs (especially so far as negation is concerned).  The algorithm to determine which tapes are atomic is actually pretty complex, and I think it's quite possible that programmers will be surprised which tapes, exactly, are treated as atomic.  And it means that the outputs of one module could change when imported into another (for example, if a tape is concatenable in my module and I import yours, it will interpret your module as concatenable, potentially changing the set of outputs from what it would have output on its own).

On the other hand, the result of ~WINTER leading to things like WUNTER is surprising in itself!  I'm not sure this is really worse.  It may be better to just continue restricting the expression of negation by the programmer -- negation is surprising either way.