Bugfixes for multichar implementation
(1) Before multichar, we could actually do constructExpr() and collectVocab() in either order (because nothing about the vocab actually affected the building of expressions).  Now that we're tokenizing within LiteralGrammar (instead of within LiteralExpr), however, we need to collectVocab() (and thus tokenize literals) before actually constructing the expression.

(2) As mentioned in the last commit, there's a bit of a confusion about namespaces because of the way they can be evaluated as if they're their last child.  This leads to some confusion when we ask things like "What tapes are relevant to namespace X" -- in some contexts we want all the tapes relevant to every child, whereas in other contexts we only want the tapes relevant to the last child.

This commit takes a first stab at refactoring this.  We don't really need the namespace to *be* its last child, we need that *references* to the namespace *refer* to its last child.  (E.g., if a namespace "NS" has three children X, Y, and Z, we want the bare reference "NS" to refer to NS.Z.  Previously, we had "NS" refer to NS itself, and then have it evaluate as if it's Z.)

So now, getSymbol() and resolveName() both return the last child of a namespace, rather than the namespace itself.  They do this with a convenience method getDefaultChild().  For a namespace, this returns the last child (if there are any children, otherwise it returns an EpsilonGrammar).  For a non-namespace, this just returns "this".

(3) In the last commit, I alluded to it being a mistake that NamespaceGrammars can nest inside grammars that express Exprs (e.g. SequenceGrammars).  We never actually do this, and if we did we'd potentially get some weird effects, like that the tapes relevant to a namespace aren't always the same as the tapes relevant to its last child (which is what it would be evaluated as).  (Of course, we could fix this, but again, it's not a "feature" we actually use or want to use.)

I experimented with making it so that NamespaceGrammars and grammars that express expressions be different types, such that the former can never nest in the latter.  This actually got a little thorny when it came to TST.toGrammar compilation, however, because a lot of operations are only well-defined on the latter, but a given TstEnclosure might represent either, so it's tricky to get the type declarations of all the methods to be compatible.

In the end, I rejected this retyping (I just wanted to mention it in here, for future reference).  The reason is, at some point it would be nice to have user-defined namespace beyond Projects and Sheets (e.g. a "namespace:" op).  (This would, for example, let us compile multi-sheet projects into single-sheet projects, which would be convenient for front-end devs, without changing the logical structure of the naming.)  But if it's an op, the programmer *could* nest it inside an expression -- we could say that's syntactically/semantically erroneous to do so, but in order to detect the error and notify the programmer, we have to be able to build the erroneous tree in the first place.  So it's ultimately not appropriate to force this nesting structure in the actual type system of our compiler code.