Generation option object + multichar/rename bugfix
(1) Now that we're getting into more optimizations, it was annoying that we couldn't turn them on and off for testing, comparison, etc.  Rather than add a new boolean parameter to compilation/generation methods every time we add a new optimization, there's now a GenOptions object that holds them.  (It also now holds maxRecusion and maxChars.)

This is mostly at the Grammars level right now, the other levels still use parameters, but eventually we should probably switch the other levels to this, too.  (Not the Gramble edifice object, though.  The whole point of that one is so clients don't have to understand our object model; that one should always just stay as simple parameters.)

(2) Negation tests were failing because of how multichar changed how characters were counted; now we turn multichar off for the purposes of those tests.  (This is better than changing the count numbers on those tests, because the multichar optimization will probably change in the future, too, meaning that these counts aren't final.  Better that tests that rely on character counting just don't turn on that optimization in the first place.)

(3) The multichar experiments revealed (or really, highlighted an old problem) that one of the behaviors of RenamedTapes was inconsistent with how tapes work everywhere else.

It's the old question of whether Rename(t1:ab+t2:cd, t1, t2) should be {t2:abcd} or should it be {} or maybe even {t2:ab}.  We've gone back and forth about this one.  I'd preferred that it have *some* output, so when it happened to give {t2:abcd} I thought this was fine.  However, this relies on the fact that RenamedTape.matchTape(tapeName) doesn't check whether tapeName is equal to its .toTape, it just forwards on the match request to its child.  (This means that when we're inside the rename state, and the t2:b needs to match on t2, matchTape("t2") still works even though t2 appears as t1 in that scope.)  That seems all well and good, but elsewhere in the system, tape renaming doesn't allow this because of various bugs that popped up.

So this reared its head again in the case of multichar detection.  The algorithm for detecting and flagging single-char sequences vs. multichars was using the "no passthru in case of tapeName==toTape" rule, so it wasn't marking t2:cd as concatenable.  Even if it were, it's not clear that we can straightforwardly determine that it's concatenable in the first place, because at the time it concatenates it's concatenating t1 to t2!  We could probably work this out (it's a similar problem to other renaming and concat problems), but it'd probably lead to more problems down the road.  The more fundamental problem is just that RenamedTape and everything else have a slightly different semantics of renaming, and so long as they do, problems like this will probably keep occuring.

So this commit makes RenamedTape.matchTape work like everything else, and fail when matching a tapeName that is equal to .toTape.  This changes the results of two tests, making them have null results.