Making Tape.inVocab work with multichars
Tape.inVocab was still assuming tokens could only contain single characters; now it works with multichar tokens in the vocab as well.

This doesn't actually fix any of the failing tests (the problem is deeper there, we'll need a real ReplaceGrammar object to get replacement working with multichar) but it's a prerequisite for future fixes.