MemoExpr limits; turned off automatic unit testing
For unclear reasons, MemoExprs no longer cause the slowdown they did a few weeks ago -- they are working fine now even with a larger grammar than then.  (There's still not a reliable speedup in the browser, though, at least not for the grammar/conditions I'm testing.)

Nonetheless, it's clear that at SOME size of grammar they will exceed the available memory, for a diminishing return, so in this commit I set a limit on how "deep" the memoization can go (very much like the limit on CompileState).  In MemoExpr, the successor to a MemoExpr is also a MemoExpr (unless it's something trivial like an epsilon or null), so we're effectively building a tree of exprs.  By analogy to CompileState, the return-on-investment of memoizing those first few generations of the tree is probably much higher than the later ones -- the bigger the tree grows, we are investing exponentially more memory for a diminishing return.  So now we can parameterize MemoExpr w.r.t. how deep that tree can get.

(Although, at the moment, we don't have memoization turned on at all because it doesn't really seem to matter for our current big grammar.)

This commit also turns off running unit tests automatically, since they can really take up a lot of time at the point of grammar-syncing.  This also accidentally turns off some important checking of regular errors, so in a near-future commit I'll separate those out, run the regular error checking automatically, and put the running of unit tests in the hands of the dev instead.