Encapsulating replacement into ReplaceGrammar
Unlike other kinds of operations, we were still treating Replacement as constructing itself out of Grammar convenience functions (e.g. Seq) rather than creating a ReplaceGrammar object and constructing its expression in constructExpr() using constructX() smart constructors.  This commit introduces a ReplaceGrammar object that does exactly that.

The resulting algorithm is a bit more verbose (it was convenient being able to just query one of the Grammar subcomponents about its tapes, which we can't do anymore), but this is ultimately necessary because we need an object in the Grammar hierarchy representing this operation.  Eventually we can take a look at refactoring, but for now I wanted to keep things one-to-one between the old code and the new so Darlene and I have a common reference.