Run unit tests from GSuite toolbar
In a previous commit, I disabled the automatic running of unit tests every time a project was loaded/synced, because (as grammars and number of tests grow) that can get prohibitively long.  But this left no way of running unit tests in the interface.  This commit adds a simple "Run unit tests" command to the GSuite toolbar.