Fixing grammar-level error checks
Grammar-level errors (e.g. undefined symbols and inappropriately-taped hides) weren't appropriately hooked up to the interface, because they were happening at the time of expression generation, and that's later than syntax highlighting in the interface.  This commit puts them all in runChecks(), and then calls that in the interface before syntax highlighting occurs.

This commit also adds two grammar-level error checks to RenameGrammar:

(a) It's now erroneous to rename a missing tape, analogously to how it's erroneous to hide a missing tape.

(b) In consequence to a recent semantic change to renaming, it's now also erroneous to rename a tape to a tape that already exists in the scope of the rename.  (E.g., if there's a tape "unit" within the scope of the rename, you can't rename the "text" tape to "unit".)