Support for named parameters to operator
(1) Named parameters for operators like "replace:"

At least in the modern incarnations of Gramble, it happened that there were never any operators that had more than 2 parameters, so the binary tree of operators and tables was sufficient.  (I.e., an op's first parameter was the grammar associated with the op above it, and the second parameter was the grammar associated with the op or table to its right.)

That's not true of replacement, replacement has at least 5 parameters (the grammar acted upon, the "from" material, the "to" material, and "before"/"after" material).  So we have to have some means, in the tabular syntax, of identifying those parameters.

Now, by prepending a header with a parameter name like "to", "before", etc., you can tag it as a particular parameter of the op:

    replace:, text, to text, after text, before text

The way this works is that some headers have an additional property, a "tag".  Most headers just contribute the default tag, defined to be "__".  Rather than just asking for toGrammar()-type methods, we now ask for toParams(), which returns a { tag: grammar} dictionary.  Tables have a special method toParamsTable(), which returns a list of these.  (Tables used to be exclusively associated with alternations, but returning an AlternationGrammar doesn't work here, because the children of grammars are grammars, and each row now represents {tag:grammar}.  So instead we just return a list of {tag:grammars} associated with each row, and for good measure the row-cell that the row is associated with, because we need that for unit test reporting.)

When an assignment, operator like "join", etc. is expecting a grammar, you can still call toGrammar().  That's now just a wrapper around toParamsTable() that marks an error if it finds anything except the default tag "__".  But an operator like "replace" can call toParamsTable() and then consume the resulting tagged grammars appropriately.

One important thing to note about tags as currently implemented is that they're transparent to each other (as they should be, we don't want to concatenate "from" and "to") but that they CAN be concatenated with other columns with the same tag.  For example, you can absolutely do something like this, meaning "turn <b> to <v> after <VOWEL>+r".

VOWEL:, text
      , a
      , i
      , u

replace:, text, to text, after embed, after text,
        , b    , v     , VOWEL      , r

(2) Changes to semantics of empty cells

There's also an important semantic change that falls out from all of this (although not one that happens to change any of our tests).  It used to be that empty cells were (from the point of view of Grammars) non-existent.  They had associated TSTs, so that they could be appropriately colored in the interface, but no actual grammar was made corresponding to them.

This was fine when the primary semantics of table rows was concatenation.  Epsilon grammars, empty literals, and not-making-a-grammar-at-all are all identical with respect to concatenation, so we might as well just not make that thing into a grammar.

(This was questionable when we started allowing rows to also express filters, but I never really had a sense of what would be better, so I just kept on doing it the old way.)

However, now that tables can express named parameters to arbitrary functions, it's not necessarily the case that these are all the same.  For a given parameter to a function, passing in an epsilon, passing in an empty literal, and not passing in that parameter might have different or strange results.  For example, it's an ordinary kind of replacement rule to replace some character <c> with nothing.  But we don't pass in anything to the "to" parameter of replacement, or pass in an epsilon, the replacement operator actually doesn't have enough information -- it doesn't know what tape to use as the output tape.

So, empty cells now have a different default semantics -- as empty literals on the relevant tape, rather than not becoming grammars at all.

This changes the semantics of filters a bit, since not filtering, and filtering such that the contents of tape T are "", are completely different.  (It happens we weren't testing for this, because I wasn't completely settled what we wanted this to really do.)  But the new behavior is probably a good thing overall.  We had ways to not filter... but we didn't have a way to filter for a particular tape being empty!  Now we do.  (Note to self to write some tests for this behavior.)