Hooking up replacement to table syntax
Replacement is now (at least in its basic form) integrated with the table syntax, so that you can go:

  table:, text
        , foo

  replace:, from text, to surface
          , o        , a

There are a bunch of little things to fix, though, for example it doesn't allow from/to the same tape yet (via renaming), we still have a maxExtraChars of 2 (we'll need infinite), etc.

Changes that were necessary to get these hooked up:

(1) fromTapeName and toTapeName are no longer arguments of Replace or new ReplaceGrammar; they're read off the fromTape and toTape arguments.

(2) In practice, when we say "to X", the X tape needs to contain the vocabulary of the Y tape of "from Y" -- otherwise, it only has the vocabulary of the cells directly under "to X".  I added a "vocabBypass" parameter that, if set to true, copies the vocabulary of Y to X in this situation, and then doesn't bother to check it when constructing the replacement expr.  Note that we can't do this at vocabulary collection time -- it doesn't need the vocabulary of Y *at the time of collection* because that might be incomplete, it needs the *final* vocabulary.  So this has to be an additional step after collectVocab.