Infinite maxExtraChars for tabular replace
Every time we construct a replace from the tabular syntax, it's in the scope of a join, so it's not as dangerous to have an infinite maxExtraChars parameter.  So the default for the ReplaceGrammar constructor is now maxExtraChars=Infinity.  (However, the Replace() convenience function still has its existing maxExtraChars=100.)

Also added tests in testHeaders for from/to/pre/post, and tests in testParams for pre/post contexts.

Also slightly refactored from/to/pre/post headers so that they're all the same type (TagHeader) rather than each having its own type.  (Don't want to have to define a new type every single time we add a named parameter to an op.)